feat: implement Spring Security with JWT and base architecture adjustments

+++ src/main/java/upc/ecovolt/mapping/dto/subscriptionplandto/SubscriptionPlanResponseDto.java
@@ -1,18 +1,18 @@
 package upc.ecovolt.mapping.dto.subscriptionplandto;
 
 import lombok.Data;
-
 import java.math.BigDecimal;
 
 @Data
 public class SubscriptionPlanResponseDto {
-
-    //Lo que al usuario le va a llegar
-    private Long id;
+    private Integer id;
     private String name;
     private BigDecimal monthlyPrice;
     private Integer deviceLimit;
+
+    /* Aquí el usuario verá el texto (ej: "Premium", "Básico") */
     private String supportLevel;
+
     private String billingCycle;
     private Integer status;
 }
